% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gBenchmark.R
\name{benchmark_bnd}
\alias{benchmark_bnd}
\title{benchmark_bnd}
\usage{
benchmark_bnd(
  x,
  y,
  ignore.strand = FALSE,
  max.dist = 1000,
  genome = "hg19",
  std.only = TRUE,
  verbose = FALSE
)
}
\arguments{
\item{x}{character vector containing a file path, GRangesList, gGraph, GRanges, or Junction object (sv calls)}

\item{y}{character vector containing a file path, GRangesList, gGraph, GRanges, or Junction object (ground truth)}

\item{ignore.strand}{(logical) perform unsigned comparison of breakends? default FALSE}

\item{max.dist}{(numeric) maximum distance between breakends to be considered a match, default 1 Kbp}

\item{genome}{(character) one of hg19, hg3. default hg19}

\item{std.only}{(logical) standard chromosomes only? default TRUE}

\item{verbose}{(logical) message stuff? default FALSE}
}
\description{
compare two sets of breakend calls
}
\details{
given two sets of breakend calls x and y, with y being the 'ground truth' or 'gold standard'
calculate precision, recall, and f1 score of x
}
