% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gBenchmark.R
\name{benchmark_cn}
\alias{benchmark_cn}
\title{benchmark_cn}
\usage{
benchmark_cn(
  x,
  y,
  x.field = "cn",
  y.field = "cn",
  allelic = FALSE,
  cnloh = FALSE,
  tile.width = 10000,
  min.width = 0,
  verbose
)
}
\arguments{
\item{x}{character vector containing a file path, data.table, GRanges, or gGraph}

\item{y}{character vector containing a file path, data.table, GRanges, or gGraph}

\item{x.field}{(character) field(s) in x containing CN data (default "cn")}

\item{y.field}{(character) field(s) in y containing CN data (default "cn")}

\item{allelic}{(logical) allele analysis? default FALSE}

\item{cnloh}{(logical) cnloh analysis? default FALSE}

\item{tile.width}{(numeric) tile width for averaging copy number estimates (default 10 Kbp)}

\item{min.width}{(numeric) minimum width segment to retain (default 0, retains all segments)}

\item{verbose}{(logical) default FALSE}
}
\description{
compare copy number (CN) predictions
}
\details{
checks consistency between total CN, allelic CN, or CNLOH predictions

inputs can be one of the follow objects, or a file containing one of the following objects:
- bigwig
- data.table coercible to GRanges
- GRanges
- gGnome gGraph
}
