% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gBenchmark.R
\name{grab_chrom_sizes}
\alias{grab_chrom_sizes}
\title{grab_chrom_sizes}
\usage{
grab_chrom_sizes(genome, std.only = TRUE)
}
\arguments{
\item{genome}{(character) hg19 or hg38?}

\item{std.only}{(logical) only "standard" contigs}
}
\value{
named numeric vector
}
\description{
helper function to get seqlengths
}
