% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gBenchmark.R
\name{read_bnds}
\alias{read_bnds}
\title{read_bnds}
\usage{
read_bnds(x, field = NA_character_, ignore.strand = FALSE)
}
\arguments{
\item{x}{character vector containing file path, or GRanges, gGraph, Junction}

\item{field}{(character) copy number field, default NA}

\item{ignore.strand}{(logical) return output with no strand information (default FALSE)}
}
\value{
GRanges (possibly unsigned) with breakend locations
}
\description{
grab breakends from input file or object
}
