% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gBenchmark.R
\name{read_segs}
\alias{read_segs}
\title{read_segs}
\usage{
read_segs(x, field = "cn", allelic = FALSE, cnloh = FALSE, xy.sub = TRUE)
}
\arguments{
\item{x}{character vector containing a file path, data.table, GRanges, bigWig, or gGraph}

\item{field}{field(s) in x containing copy number data (default 'cn')}

\item{allelic}{(logical) allelic analysis? default FALSE}

\item{cnloh}{(logical) CNLOH analysis? default FALSE}

\item{xy.sub}{(logical) if '23' and '24' are provided as seqnames, replace them with X and Y? default TRUE}
}
\value{
GRanges with field "score"
score represents copy number if "cnloh" is TRUE
additional field "allele" with values "major" and "minor" if "allelic" is TRUE
}
\description{
read segments and copy number from input file
}
