% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gBenchmark.R
\name{seg2bnd}
\alias{seg2bnd}
\title{seg2bnd}
\usage{
seg2bnd(gr)
}
\arguments{
\item{gr}{}
}
\value{
signed GRanges representing breakends
}
\description{
convert segments with 'score' field representing CN to breakends
}
