% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gBenchmark.R
\name{simulate_coverage}
\alias{simulate_coverage}
\title{simulate_coverage}
\usage{
simulate_coverage(
  gr,
  field = "cn",
  bins = NULL,
  binsize = 1000,
  diploid = TRUE,
  bias = NULL,
  bias.field = "background",
  basecov = 60,
  readsize = 150,
  purity = 0.95,
  poisson = TRUE,
  overdispersion = NA,
  normalize = TRUE,
  genome = "hg19"
)
}
\arguments{
\item{gr}{(GRanges, gGraph, or file containing one of these things) segment copy number, with numeric cn supplied in field "cn"}

\item{field}{(character) field containing copy number in gr, default cn}

\item{bins}{(GRanges) genomic bins}

\item{binsize}{(numeric) width of genomic bins (bp) for simulating read counts. ignored if bins if supplied. default 1e3.}

\item{diploid}{(logical) simulate diploid genome? default TRUE}

\item{bias}{(GRanges) mulitplicative bias for binned means}

\item{bias.field}{(character) column in bias GRanges metadata containing multiplicative bias. default "background"}

\item{basecov}{(numeric) expected coverage depth. default 60.}

\item{readsize}{(numeric) expected read size (bp). default 150.}

\item{purity}{(numeric) sample purity, between zero and one. default 0.95.}

\item{poisson}{(numeric) simulate actual integer reads? if NA will just output relative copy number.}

\item{overdispersion}{(numeric) if supplied, should be > 1 and will be applied to simulate from a negative binomial distribution. default NA.}

\item{normalize}{(logical) unit normalize resulting reads by dividing by mean? default TRUE}
}
\value{
GRanges with coverage per genomic bin in field "cov"
}
\description{
simulate binned read counts given expected depth
}
\details{
Given a genomic segments with estimated tumor copy number (CN) and genomic bin size, simulate read counts per genomic bin.
This can be done for either a diploid genome (diploid = TRUE) or haploid genome (haploid = TRUE).

Optionally, a multiplicative bias (representing replication timing, batch effect, etc.) can be supplied to make the coverage appear more realistic.
This should be supplied as a GRanges as a numeric vector in field "background" (or some other field specified in bias.field)

The expected coverage and read size can be tuned by setting basecov and readsize, respectively.

By default, read counts are simulated from a Poisson distribution, but if overdispersion (numeric) is supplied will simulate from a Negative Binomial distribution.
}
