% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gBenchmark.R
\name{benchmark_bnd}
\alias{benchmark_bnd}
\title{benchmark_bnd}
\usage{
benchmark_bnd(
  x,
  y,
  x.field = NA_character_,
  y.field = NA_character_,
  ignore.strand = FALSE,
  max.dist = 1000,
  genome = "hg19",
  std.only = TRUE,
  mask = system.file("extdata", "hg19.mask.rds", package = "gBenchmark"),
  verbose = FALSE
)
}
\arguments{
\item{x}{character vector containing a file path, GRangesList, gGraph, GRanges, or Junction object (sv calls)}

\item{y}{character vector containing a file path, GRangesList, gGraph, GRanges, or Junction object (ground truth)}

\item{x.field}{(character) default NA}

\item{y.field}{(character) default NA}

\item{ignore.strand}{(logical) perform unsigned comparison of breakends? default FALSE}

\item{max.dist}{(numeric) maximum distance between breakends to be considered a match, default 1 Kbp}

\item{genome}{(character) one of hg19, hg3. default hg19}

\item{std.only}{(logical) standard chromosomes only? default TRUE}

\item{mask}{(character) path to coverage mask}

\item{verbose}{(logical) message stuff? default FALSE}
}
\description{
compare two sets of breakend calls
}
\details{
given two sets of breakend calls x and y, with y being the 'ground truth' or 'gold standard'
calculate precision, recall, and f1 score of x

x and y can be (signed) GRangesList, GRanges, gGraph (from which junctions/loose ends are extracted), or Junctions (or a file path containing any of these objects). in each of these cases, x.field and y.field do not have to be supplied.

in addition, x and y can be a segmentation output (such as GRanges) with some field containing copy number (supplied as x.field or y.field). using the copy number change points, a set of signed breakends are inferred.

if ignore.strand == TRUE, the orientation of breakends is ignored and two breakends in potentially opposite directions may be considered overlapping.
}
\examples{
## compare a gGraph (x) with a set of junction calls provided as vcf (y)
res = gBenchmark:::benchmark_bnd(x = system.file("extdata", "gg.rds", package = "gBenchmark"),
                                y = system.file("extdata", "svaba.subset.vcf", package = "gBenchmark"))

## compare a gGraph (x) with a set of junction calls (provided as grl) (y)
res = gBenchmark:::benchmark_bnd(x = system.file("extdata", "gg.rds", package = "gBenchmark"),
                                y = system.file("extdata", "grl.rds", package = "gBenchmark"))

## compare a GRanges (x) with a set of junction calls (provided as grl) (y) and ignore the strand
res = gBenchmark:::benchmark_bnd(x = system.file("extdata", "gr.rds", package = "gBenchmark"),
                                y = system.file("extdata", "grl.rds", package = "gBenchmark"),
                                ignore.strand = TRUE)


}
